requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
fastapi==0.109.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httpx==0.26.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
mypy==1.8.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
pluggy==1.4.0; python_version >= '3.8'
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pytest==7.4.4; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.35.1; python_version >= '3.8'
types-python-dateutil==2.8.19.20240106; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
uvicorn==0.27.0; python_version >= '3.8'
